Make `decimals` be stored as `Int` instead of `UInt8`. Latter (to match actual source data in ERC20) seems unnecessary

+++ AlphaWallet/Tokens/ViewControllers/NewTokenViewController.swift
@@ -242,7 +242,7 @@ class NewTokenViewController: UIViewController {
         nameTextField.value = name
     }
 
-    public func updateDecimalsValue(_ decimals: UInt8) {
+    public func updateDecimalsValue(_ decimals: Int) {
         decimalsTextField.value = String(decimals)
     }
