+++ modules/AlphaWalletFoundation/AlphaWalletFoundation/Tokens/Logic/GetContractDecimals.swift
@@ -11,11 +11,11 @@ public class GetContractDecimals {
         self.server = server
     }
 
-    public func getDecimals(for contract: AlphaWallet.Address) -> Promise<UInt8> {
+    public func getDecimals(for contract: AlphaWallet.Address) -> Promise<Int> {
         let functionName = "decimals"
-        return callSmartContract(withServer: server, contract: contract, functionName: functionName, abiString: Web3.Utils.erc20ABI).map { dictionary -> UInt8 in
-            guard let decimalsOfUnknownType = dictionary["0"], let decimals = UInt8(String(describing: decimalsOfUnknownType)) else {
-                throw CastError(actualValue: dictionary["0"], expectedType: UInt8.self)
+        return callSmartContract(withServer: server, contract: contract, functionName: functionName, abiString: Web3.Utils.erc20ABI).map { dictionary -> Int in
+            guard let decimalsOfUnknownType = dictionary["0"], let decimals = Int(String(describing: decimalsOfUnknownType)) else {
+                throw CastError(actualValue: dictionary["0"], expectedType: Int.self)
             }
 
             return decimals
